backend/requirements.txt
Django>=5.0
psycopg[binary]
djangorestframework
paho-mqtt
